Mutable Cursors to remove resource leaks
Why:
It is strange to have a new cursor each time a read happens. This
changes that reads only mutate a single variable and keep the
existing cursor around. This is strange and perhaps not ideal.
Perhaps read should return the result from `readInternal` and set
a flag to inform the WritableCursor that it needs to seek to the
end of the file to execute it's write.