Add deleteAtTimestamp and remove delete method on cursor
Why:
Wanted to add the functionality to delete a item at a particular
timestamp. Given this it seems silly to duplicate AtTimestamp methods
this change just uses the writeAtTimestamp methods for deletes

How:
deleteAtTimestamp uses writeAtTimestamp with value of the
Tombstone. Then the delete method now uses deleteAtTimestamp to make
it all work together. This creates a stack of
delete -> deleteAtTimestamp -> writeAtTimestamp which could be
simplified to just delete -> writeAtTimestamp.